+++ gateway/src/apicast/policy/logging/logging.lua
@@ -80,6 +80,7 @@ local function get_request_context(context)
   ctx.service = context.service or {}
   ctx.original_request = context.original_request
   ctx.jwt = context.jwt or {}
+  ctx.credentials = context.credentials
   return LinkedList.readonly(ctx, ngx.var)
 end